Parse for statement correctly in one cornercase
The cornercase:
for (f = 0.0; f != 3.3; ++f)

(related to first expression, f=0.0)

+++ src/glslplugin/lang/parser/GLSLParsing.java
@@ -701,7 +701,7 @@ public final class GLSLParsing extends GLSLParsingBase {
         }
 
         PsiBuilder.Marker rollback = b.mark();
-        if(parseDeclarationStatement()){
+        if(lookaheadDeclarationStatement() && parseDeclarationStatement()){
             rollback.drop();
             return;
         }else{
